% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{acubemap_dependency}
\alias{acubemap_dependency}
\title{Dependency}
\usage{
acubemap_dependency()
}
\description{
\code{acubemap} dependency.
}
\examples{
library(aframer)

a_scene(
  a_dependency(),
  acubemap_dependency()
)

}
